finsh topic and video

+++ database/migrations/2019_06_07_062214_create_topics_table.php
@@ -18,7 +18,7 @@ class CreateTopicsTable extends Migration
             $table->string('title')->comment('话题标题');
             $table->string('front_cover')->comment('封面');
             $table->string('content')->comment('话题内容');
-            $table->tinyInteger('sort')->default(0)->comment('排序');
+            $table->integer('sort')->default(0)->comment('排序');
             $table->tinyInteger('status')->default(1)->comment('状态：0冻结、1激活');
             $table->timestamps();
         });
